@@ -3,78 +3,35 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint --fix src"
   },
   "dependencies": {
     "@quasar/extras": "1.16.11",
-    "@traptitech/traq": "3.4.5-1",
-    "axios": "1.6.8",
-    "core-js": "3.37.0",
+    "@traptitech/traq": "3.17.0-2",
     "dayjs": "1.11.10",
     "github-markdown-css": "5.5.1",
-    "quasar": "1.22.10",
-    "register-service-worker": "1.7.2",
-    "vue": "2.7.16",
-    "vue-router": "3.6.5",
-    "vuex": "3.6.2",
+    "quasar": "2.15.3",
+    "vue": "3.4.23",
+    "vue-router": "4.3.2",
+    "vuex": "4.1.0",
     "vuex-persistedstate": "3.2.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "4.5.13",
-    "@vue/cli-plugin-eslint": "4.5.13",
-    "@vue/cli-plugin-pwa": "4.5.13",
-    "@vue/cli-plugin-router": "4.5.13",
-    "@vue/cli-plugin-vuex": "4.5.13",
-    "@vue/cli-service": "4.5.13",
-    "@vue/eslint-config-standard": "6.1.0",
-    "babel-eslint": "10.1.0",
-    "babel-plugin-transform-imports": "2.0.0",
-    "eslint": "7.32.0",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "5.2.0",
-    "eslint-plugin-standard": "5.0.0",
-    "eslint-plugin-vue": "7.20.0",
+    "@quasar/vite-plugin": "1.6.0",
+    "@vitejs/plugin-vue": "5.0.4",
+    "eslint": "9.1.0",
+    "eslint-plugin-vue": "9.25.0",
     "markdown-it": "12.3.2",
     "markdown-it-mark": "3.0.1",
     "pug": "2.0.4",
     "pug-loader": "2.4.0",
     "pug-plain-loader": "1.1.0",
-    "stylus": "0.55.0",
-    "stylus-loader": "3.0.2",
-    "vue-cli-plugin-quasar": "3.0.2",
-    "vue-template-compiler": "2.7.16",
-    "ware-loader": "0.2.4"
-  },
-  "resolutions": {
-    "kind-of": "^6.0.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "@vue/standard"
-    ],
-    "rules": {
-      "vue/custom-event-name-casing": "error"
-    },
-    "parserOptions": {
-      "parser": "babel-eslint"
-    }
-  },
-  "postcss": {
-    "plugins": {
-      "autoprefixer": {}
-    }
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not ie <= 8"
-  ]
+    "stylus": "0.63.0",
+    "vite": "5.2.9",
+    "vite-plugin-markdown": "2.2.0",
+    "vite-plugin-pwa": "0.19.8"
+  }
 }
