@@ -1,5 +1,3 @@
 <template>
-  <div id="app">
-    <router-view/>
-  </div>
+  <router-view />
 </template>
